+++ changelog.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - add release script for MS
 - add release scripts for Linux
 - add release scripts for OSx
-
+- make output CSV configurable,  #11
 
 ### Fixed